fix: use latest etl procedures
In pervious update of daily/annual income, I used an outdated etl
script and updated the method based on that script. But the script
didn't include the latest changes in the etl procedure (i.e. add
chn to countries for those with hkg, twn, mac), and produced a
different population-by-income_groups datapoints. This version corrects
the problem, and produced the same population-by-income_groups datapoints
as in 6318457554563